Fix issue with zombie esbuild processes
Extracted from #322 since this bug isn't unique to that branch, it exists in
main as well.

When the dev server restarts, it successfully terminates the npm exec process.
But, the child esbuild processes don't terminate as expected. Thus, after a
short development period, quite a number of esbuild processes can pile up on the
system.

This changes fixes the issue by starting each process in a new process group.
Then, instead of terminating the npm process itself, we terminate the entire
process group.

This logic is based on the solution described here: https://stackoverflow.com/a/32222971/162220